+++ .gitignore
@@ -1,0 +1,2 @@
+
+HyperledgerAries/CHAPI_test/package-lock.json
