+++ HyperledgerAries/CHAPI_test/main.html
@@ -1,10 +1,57 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <meta charset="UTF-8">
-    <title>https://cocoon1787.tistory.com/</title>
+    <title>Credential Handler Example</title>
+    <script src="https://unpkg.com/credential-handler-polyfill@3.0.1/dist/credential-handler-polyfill.min.js"></script>
   </head>
   <body>
-    <div>환영합니다.</div>
+    <h1>Credential Handler Example</h1>
+    <p>Click the button to retrieve your credentials:</p>
+    <button onclick="getCredentials()">Get Credentials</button>
+    <script>
+      credentialHandlerPolyfill
+    .loadOnce(MEDIATOR)
+    .then(console.log('Polyfill loaded.'))
+    .catch(e => console.error('Error loading polyfill:', e));
+      
+      const testCredential = {
+    "@context": [
+        "https://www.w3.org/2018/credentials/v1",
+        "https://www.w3.org/2018/credentials/examples/v1"
+      ],
+      "id": "http://example.edu/credentials/1872",
+      "type": ["VerifiableCredential", "UniversityDegreeCredential"],
+      "issuer": "https://example.edu/issuers/565049",
+      "issuanceDate": "2010-01-01T19:53:24Z",
+      "credentialSubject": {
+        "id": "did:example:ebfeb1f712ebc6f1c276e12ec21",
+        "alumniOf": {
+          "id": "did:example:c276e12ec21ebfeb1f712ebc6f1",
+          "name": {
+            "@value": "Example University",
+            "@language": "en"
+          }
+        }
+      },
+      "proof": {
+        "type": "Ed25519Signature2018",
+        "created": "2017-06-18T21:19:10Z",
+        "proofPurpose": "assertionMethod",
+        "verificationMethod": "https://example.edu/issuers/keys/1",
+        "jws": "eyJhbGciOiJSUzI1NiIsImI2NCI6ZmFsc2UsImNyaXQiOlsiYjY0Il19..TCYt5XsITJX1CxPCT8yAV-TVkIEq_PbChOMqsLfRoPsnsgw5WEuts01mq-pQy7UJiN5mgRxD-WUcX16dUEMGlv50aqzpqh4Qktb3rk-BuQy72IFLOqV0G_zS245-kronKb78cPN25DGlcTwLtjPAYuNzVBAh4vGHSrQyHUdBBPM"
+      }
+  };
+
+  const testPresentation = {
+    "@context": [
+      "https://www.w3.org/2018/credentials/v1",
+      "https://www.w3.org/2018/credentials/examples/v1"
+    ],
+    "type": "VerifiablePresentation",
+    "verifiableCredential": [
+      testCredential
+    ]
+  };
+    </script>
   </body>
 </html>
