+++ HyperledgerAries/CHAPI_test/myscript.js
@@ -1,0 +1,5 @@
+function greet(name) {
+  console.log(`Hello, ${name}!`);
+}
+
+module.exports = greet;
