+++ HyperledgerAries/CHAPI_test/package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "description": "",
   "main": "server.js",
+  "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node server.js"
@@ -10,6 +11,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "credential-handler-polyfill": "^3.2.0"
+    "credential-handler-polyfill": "^3.2.0",
+    "express": "^4.18.2"
   }
 }
