+++ Minutes/2023_05_01.md
@@ -17,7 +17,7 @@
       - unity 상의 python 스크립드 작성 확인
 
 #### 이번주에 할일
-  - Aries 분석 (15 시간)
+  - Aries 분석 
     - Aries Cloud Agent 실행 테스트
       - 최종 VP 발급 이후 검증 작업 필요 -> indy-sdk를 사용한 검증과 같은가 확인 필요
     - Aries 코드 분석
@@ -25,11 +25,11 @@
       - Aries rfc 문서를 확인하여 각 Protocols의 목적 및 방법 분석 
         - [Protocol 디렉토리 출력](/HyperledgerAries/aries-python-test/README.md) 
       - DOC 문서 : https://docs.google.com/document/d/1tfVqI25hKeYxKjjefntZ6yZxhOSor1BSBjOJxlaUcsM/edit?usp=sharing
-  - Avata DID 부여 (설계 진행 중) (10시간)
+  - Avata DID 부여 (설계 진행 중) 
     - Avata에 DID 부여시 시스템 설계
       - unity 상의 python 스크립드 작성 확인 -> 실제 실행 시 동작 및 indy-sdk 패키지 사용 또한 필요
     - Notion에 정리 중 : https://curvy-tip-bb0.notion.site/Avata-DID-2e712308e54b449db0b5b52bda1135c9
-  - CHAPI 분석 (https://chapi.io/) (5시간)
+  - CHAPI 분석 (https://chapi.io/) 
     - Node.js를 사용한 웹 구현이 필요 -> Node.js에 자바 스크립트 사용 공부
     - [레퍼지토리](/HyperledgerAries/CHAPI_test)에 코드 공유 및 정리 중
