+++ Minutes/2023_05_08.md
@@ -1,1 +1,52 @@
+## 2023. 5. 8 회의록
 
+내용을 구체적으로 정리합시다. 한일에 재목만 적지 말고 제목과 관련해서 한일을 구체적으로 그리고 남은 일도 함께 적읍시다. 그리고 결정해야 할 내용도 함께 작성.
+
+
+### 최규현
+
+#### 지난주에 한일 (26시간)
+  - Aries 분석
+    - Aries Cloud Agent 실행 테스트
+      - 최종 VP 발급 이후 검증 확인 -> 확인 완료 ("verified": "true")
+    - Aries 보고서 작성
+      - 현재까지 공부한 Aries 내용 정리 및 작성
+        - VC, VP 내용 정리 중 -> 각 속성 및 의미
+  - Avata DID 부여
+    - Avata에 DID 부여시 시스템 설계
+      - unity 상에서 'indy-sdk' 실행 -> 지갑 생성 및 풀 연결 가능 -> 2번 이상 실행할 경우 문제 발생 -> 확인 필요
+      - [레퍼지토리](/HyperledgerAries/Avata DID)에 정리 중
+  - CHAPI 분석 (https://chapi.io/) (5시간)
+    - Node.js 사용해 웹 사이트 구현 -> javaScript 파일 연동 및 HTML 파일 사용 -> CHAPI 패키지 사용에 문제 발생 -> 확인 필요
+
+
+#### 이번주에 할일
+  - Aries 분석 
+    - Aries Cloud Agent 실행 테스트
+      - 로그 확인을 통
+    - Aries 코드 분석
+      - Aries rfc 문서를 확인하여 각 Protocols의 목적 및 방법 분석 
+        - [Protocol 디렉토리 출력](/HyperledgerAries/aries-python-test/README.md) 
+      - DOC 문서 : https://docs.google.com/document/d/1tfVqI25hKeYxKjjefntZ6yZxhOSor1BSBjOJxlaUcsM/edit?usp=sharing
+  - Avata DID 부여
+    - Avata에 DID 부여시 시스템 설계
+      - python의 비동기 함수 확인 필요 (indy-sdk python이 비동기로 작동 중) -> async, await 함수 및 asyncio 패키지
+      - [레퍼지토리](/HyperledgerAries/Avata DID)에 정리 중
+  - CHAPI 분석 (https://chapi.io/) 
+    - Node.js에서 CHAPI 패키지 사용을 위한 정보 확인 -> import 문제 해결이 필요
+    - [레퍼지토리](/HyperledgerAries/CHAPI_test)에 코드 공유 및 정리 중
+
+### 홍유성
+
+#### 지난주에 한일 (10시간)
+
+
+#### 이번주에 할일
+
+
+### 김동우
+
+#### 지난주에 한일 (10시간)
+
+
+#### 이번주에 할일
